unit_tests: In TestDerEncodeDecode, cast Object to Integer before casting to int (found by mvn compile). Refs #2374.

+++ tests/src/net/named_data/jndn/tests/unit_tests/TestDerEncodeDecode.java
@@ -249,10 +249,10 @@ public class TestDerEncodeDecode {
        name, ((Blob)decodedName.toVal()).toString());
     assertEquals
       ("Wrong extension value trust class after decoding",
-       trustClass, (int)decodedTrustClass.toVal());
+       trustClass, (int)(Integer)decodedTrustClass.toVal());
     assertEquals
       ("Wrong extension value trust level after decoding",
-       trustLevel, (int)decodedTrustLevel.toVal());
+       trustLevel, (int)(Integer)decodedTrustLevel.toVal());
 }
 
   @Test
